Начал писать все для авторов. Сделал Author, AuthorTrans, AuthorDao. Начал писать AuthorService.

+++ src/main/java/com/igorrogachev/athenaeum/dao/BookDao.java
@@ -3,5 +3,5 @@ package com.igorrogachev.athenaeum.dao;
 import com.igorrogachev.athenaeum.entity.Book;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookDao extends CrudRepository<Book, Integer> {
+public interface BookDao extends CrudRepository<Book, Long> {
 }
